Extract tests for isIntlNumberFormat to own file
